requirements1.txt
chroma-hnswlib==0.7.3
chromadb==0.5.0
faiss-cpu==1.8.0.post1
Flask==3.0.3
Flask-Cors==4.0.1
groq==0.10.0
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.35
langchain-groq==0.1.9
langchain-openai==0.1.10
langchain-text-splitters==0.2.1
nltk==3.8.1
pandas==2.2.2
pdf2image
pdfminer.six
pdfplumber
pikepdf==9.0.0
pillow==10.3.0
pyngrok==7.2.0
pypandoc==1.13
pyparsing==3.1.2
pypdf==4.2.0
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.1.0
PyStemmer==2.2.0.1
pytesseract==0.3.10
python-docx==1.1.2
python-dotenv==1.0.1
tiktoken==0.7.0
timm==1.0.3
tokenizers==0.19.1
tqdm==4.66.4